PBI-2947 Set operations for targeting (#224)
* PBI-2948 Create system for storing IRT repsonses
[PBI-2948]

* clean up record function

* Update checks.
Update tests.
Still having some serializing issues I'm looking into and fixing and tests

* Fix tests

* PBI-2348 Record Notes responses for targeting
[PBI-2348]

* Update interaction stream for interaction responses

* Fix the responses parameter

* Fix test

* ktlint

* Fix tests

* PBI-2439 Record survey responses for targeting
[PBI-2439]

* feature/PBI-2347 Expose responses to targeting
[PBI-2347]

* PBI-2947 Set operations for targeting
[PBI-2947]

* Fix some operator conditions

* Add other response type

* Add other response

* Merge and fix tests

* Fix comments

* Add other conditions and fix tests

* Rename serializer

* Specify return type

* Change response type from String to Any (Can be string or Long)

* Change id to Any type

+++ apptentive-feedback/src/main/java/apptentive/com/android/feedback/engagement/criteria/DefaultTargetingState.kt
@@ -92,7 +92,7 @@ internal data class DefaultTargetingState(
             is device.locale_language_code -> device.localeLanguageCode
             is device.locale_raw -> device.localeRaw
             is device.os_api_level -> device.osApiLevel
-            is device.utc_offset -> device.utcOffset
+            is device.utc_offset -> device.utcOffset.toLong()
             is device.custom_data -> device.customData[field.key]
 
             else -> null
